--- src/main/java/bo/usfx/exampletdd/package-info.java
@@ -1,1 +1,0 @@
-package bo.usfx.exampletdd;
