--- src/main/java/bo/usfx/package-info.java
@@ -1,1 +1,0 @@
-package bo.usfx;
