+++ src/main/java/bo/usfx/simonchumacero/assignment2/Kata1.java
@@ -1,29 +1,29 @@
-package simon.chumacero.espada.assigment2;
+package bo.usfx.simonchumacero.assignment2;
 
 public final class Kata1 {
   /**
-   * Apartir de esta cantidad se considera una serie.
-   * */
+   * From this quantity onward, it is considered a series.
+   */
   private static final int SMELL_SERIE = 3;
   /**
-   * Con una cantidad entre 1 y 2, no se considera una serie.
-   * */
+   * With a quantity between 1 and 2, it is not considered a series.
+   */
   private static final int MIN_NUMBER = 1;
   /**
-   * Con una cantidad entre 1 y 2, no se considera una serie.
-   * */
+   * With a quantity between 1 and 2, it is not considered a series.
+   */
   private static final int MAX_NUMBER = 2;
-
 
   private Kata1() {
 
   }
 
   /**
-   * Metodo que verifica si una serie dada esta correcta para ser publicada.
-   * @param x Recibe una lista de strings con una serie
-   * @return Devuelve un string con la clasificación de la serie
-   * */
+   * Method that checks if a given series is correct to be published.
+   *
+   * @param x Receives a list of strings with a series
+   * @return Returns a string with the classification of the series
+   */
   public static String well(final String[] x) {
 
     int count = 0;
