+++ src/main/java/bo/usfx/simonchumacero/assignment2/Kata2.java
@@ -1,0 +1,18 @@
+package bo.usfx.simonchumacero.assignment2;
+
+public final class Kata2 {
+
+  private Kata2() {
+
+  }
+
+  public static int summation(final int n) {
+    int sum = 0;
+
+    for (int i = 1; i <= n; i++) {
+      sum += i;
+    }
+
+    return sum;
+  }
+}
