+++ src/main/java/bo/usfx/simonchumacero/assignment3/Kata2.java
@@ -1,25 +1,18 @@
-package simon.chumacero.espada.assigment3;
+package bo.usfx.simonchumacero.assignment3;
 
 public final class Kata2 {
 
   /**
    * Constante de la conjetura Collatz.
-   * */
+   */
   private static final int CONSTANT_3 = 3;
-
-
 
 
   private Kata2() {
 
   }
 
-  /**
-   * Calcula el numero de pasos en la secuencia para un numero dado.
-   *
-   * @param n es el numero de pasos para el que se calculará la secuencia
-   * @return Número de pasos en la secuencia
-   * */
+
   public static int hotpo(final int n) {
     int steps = 0;
     int currentNumber = n;
