--- src/main/java/simon/chumacero/espada/assigment2/package-info.java
@@ -1,1 +1,0 @@
-package simon.chumacero.espada.assigment2;
