--- src/main/java/simon/chumacero/espada/package-info.java
@@ -1,1 +1,0 @@
-package simon.chumacero.espada;
